refactor: Move shared logic into a concern for enrichment extractions
